requirements.txt
# Core harness dependencies
pyyaml>=6.0.0
requests>=2.31.0

# Dashboard dependencies
streamlit>=1.28.0
plotly>=5.18.0
pandas>=2.0.0

# HR Simulator dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
